removed duct tape, deleted unused files, put icons and images into folders within assets, removed some white space, console logs, and unused imports
