+++ client/src/App.css
@@ -4,7 +4,7 @@
   font-weight: 400;
   margin: 0 auto;
   text-align: center;
-  background-image: url(./assets/wind-turbines.jpg);
+  background-image: url(./assets/images/wind-turbines.jpg);
   background-size: cover;
   background-position: center;
 }
