+++ client/src/assets/icons/earth_icon.png
