+++ client/src/assets/images/wind-turbines.jpg
