+++ client/src/pages/Profile.jsx
@@ -2,7 +2,6 @@ import { Navigate, useParams } from 'react-router-dom';
 import { useQuery } from '@apollo/client';
 import { QUERY_ME } from '../utils/queries';
 import Auth from '../utils/auth';
-import AllHistory from '../components/charts/AllHistory'
 import CarbonSnapshot from '../components/charts/CarbonSnapshot';
 import { Container, Row } from 'reactstrap';
